\name{alldescriptions-methods}
\docType{methods}
\alias{alldescriptions}
\alias{alldescriptions-methods}
\alias{alldescriptions,Dataset-method}
\title{Get the description label of all variables}
\description{
\code{alldescriptions} returns the description label for all variables of an object storing variables. The value is a named character vector.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "Dataset")}}{
\code{alldescriptions} method for a \code{Dataset} object.
}
}}
\keyword{methods}
\keyword{description}
\keyword{label}
