\name{checkvars-methods}
\docType{methods}
\alias{checkvars}
\alias{checkvars-methods}
\alias{checkvars,Dataset-method}
\title{Get the control variables used}
\description{
\code{checkvars} returns the names of the variables used for controlling the representativeness when filtering out cases.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "Dataset")}}{
\code{checkvars} method for a \code{Dataset} object.
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
