\name{description-methods}
\docType{methods}
\alias{description}
\alias{description-methods}
\alias{description,Dataset-method}
\alias{description,Variable-method}
\title{Get the description of an object}
\description{
\code{description} returns the value stored in the description attribute of an object.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "Dataset")}}{
\code{description} method for a \code{Dataset} object. Value is a \code{character}.
}

\item{\code{signature(object = "Variable")}}{
\code{description} method for a \code{Variable} object. Value is a \code{character}.
}
}}
\section{Examples}{
%data(dds)
%description(dds$sexe)
}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
