\name{weighting-methods}
\docType{methods}
\alias{weighting}
\alias{weighting-methods}
\alias{weighting,Bivan-method}
\alias{weighting,Dataset-method}
\title{Get the variable used as weighting}
\description{
\code{weighting} return the name of the variable used as weighting.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "Bivan")}}{
\code{weighting} method for a \code{Bivan} object.
}

\item{\code{signature(object = "Dataset")}}{
\code{weighting} method for a \code{Dataset} object.
}
}}
\keyword{methods}
\keyword{weights}
